+++ daangnMarket/src/main/java/org/sopt/daangnMarket/exception/GlobalExceptionHandler.java
@@ -2,6 +2,7 @@ package org.sopt.daangnMarket.exception;
 
 import org.sopt.daangnMarket.util.ApiResponse;
 import org.sopt.daangnMarket.util.ApiUtils;
+import org.sopt.daangnMarket.util.dto.ErrorMessage;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -20,19 +21,19 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(NotFoundException.class)
     public ResponseEntity<ApiResponse<String>> handleNotFoundException(NotFoundException e) {
-        ApiErrorCode errorCode = e.getErrorCode();
-        return ApiUtils.error(HttpStatus.valueOf(errorCode.getStatus()), errorCode);
+        ErrorMessage errorMessage = e.getErrorMessage();
+        return ApiUtils.error(HttpStatus.valueOf(errorMessage.getStatus()), errorMessage);
     }
 
     @ExceptionHandler(DuplicateMemberException.class)
     public ResponseEntity<ApiResponse<String>> handleDuplicateMemberException(DuplicateMemberException e) {
-        ApiErrorCode errorCode = e.getErrorCode();
-        return ApiUtils.error(HttpStatus.valueOf(errorCode.getStatus()), errorCode);
+        ErrorMessage errorMessage = e.getErrorMessage();
+        return ApiUtils.error(HttpStatus.valueOf(errorMessage.getStatus()), errorMessage);
     }
 
     @ExceptionHandler(ConflictException.class)
     public ResponseEntity<ApiResponse<String>> handleConflictException(ConflictException e) {
-        ApiErrorCode errorCode = e.getErrorCode();
-        return ApiUtils.error(HttpStatus.valueOf(errorCode.getStatus()), errorCode);
+        ErrorMessage errorMessage = e.getErrorMessage();
+        return ApiUtils.error(HttpStatus.valueOf(errorMessage.getStatus()), errorMessage);
     }
 }
