+++ daangnMarket/src/main/java/org/sopt/daangnMarket/util/dto/ErrorMessage.java
@@ -1,8 +1,12 @@
-package org.sopt.daangnMarket.exception;
+package org.sopt.daangnMarket.util.dto;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
 import org.springframework.http.HttpStatus;
 
-public enum ApiErrorCode {
+@Getter
+@AllArgsConstructor
+public enum ErrorMessage {
 
     MEMBER_NOT_FOUND(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 사용자가 존재하지 않습니다."),
     CATEGORY_NOT_FOUND(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 카테고리가 존재하지 않습니다."),
@@ -15,19 +19,6 @@ public enum ApiErrorCode {
 
     INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR.value(), "Internal server error");
 
-    private int status;
-    private String message;
-
-    ApiErrorCode(int status, String message) {
-        this.status = status;
-        this.message = message;
-    }
-
-    public int getStatus() {
-        return status;
-    }
-
-    public String getMessage() {
-        return message;
-    }
+    private final int status;
+    private final String message;
 }
